Move recon_vis out of reconcile
