+++ empress/recon_vis/recon_viewer.py
@@ -5,10 +5,9 @@ View a single reconciliation using matplotlib
 from typing import Union
 from matplotlib import pyplot as plt
 
-from empress.reconcile.recon_vis.recon import EventType
-from empress.reconcile.recon_vis import utils
-from empress.reconcile.recon_vis import plot_tools
-from empress.reconcile.recon_vis.render_settings import LEAF_NODE_COLOR, COSPECIATION_NODE_COLOR, \
+from empress.recon_vis.recon import EventType
+from empress.recon_vis import utils, plot_tools
+from empress.recon_vis import LEAF_NODE_COLOR, COSPECIATION_NODE_COLOR, \
     DUPLICATION_NODE_COLOR, TRANSFER_NODE_COLOR, HOST_NODE_COLOR, HOST_EDGE_COLOR, \
     PARASITE_EDGE_COLOR, VERTICAL_OFFSET, COSPECIATION_OFFSET
