+++ test/reconcile/recon_vis/test_recon_viewer.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import shutil
 import unittest
 
-from empress.reconcile.recon_vis import recon_viewer
+from empress.recon_vis import recon_viewer
 
 
 class TestReconViewer(unittest.TestCase):
